# Optimizes `yaw_quat` implementation